[core] Move CrossFadedPropertyEvaluator to its own file
