+++ src/mbgl/programs/fill_program.cpp
@@ -1,6 +1,6 @@
 #include <mbgl/programs/fill_program.hpp>
 #include <mbgl/sprite/sprite_atlas.hpp>
-#include <mbgl/style/property_evaluator.hpp>
+#include <mbgl/style/cross_faded_property_evaluator.hpp>
 #include <mbgl/tile/tile_id.hpp>
 #include <mbgl/map/transform_state.hpp>
